@@ -14,16 +14,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.2.16",
+    "@angular/animations": "8.2.14",
     "@angular/cdk": "7.3.7",
-    "@angular/common": "7.2.16",
-    "@angular/compiler": "7.2.16",
-    "@angular/core": "7.2.16",
-    "@angular/forms": "7.2.16",
-    "@angular/http": "7.2.16",
-    "@angular/platform-browser": "7.2.16",
-    "@angular/platform-browser-dynamic": "7.2.16",
-    "@angular/router": "7.2.16",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "8.2.14",
+    "@angular/forms": "8.2.14",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/router": "8.2.14",
     "@ngx-translate/core": "^11.0.1",
     "@ngx-translate/http-loader": "^4.0.0",
     "@openecoe/potion-client": "^4.0.12",
@@ -35,19 +34,19 @@
     "ngx-bar-rating": "^1.1.0",
     "ngx-papaparse": "^3.0.2",
     "reflect-metadata": "^0.1.13",
-    "rxjs": "~6.3.3",
+    "rxjs": "~6.6.7",
     "socket.io-client": "^2.3.0",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.12",
-    "@angular/cli": "7.3.0",
-    "@angular/compiler-cli": "7.2.16",
-    "@angular/language-service": "7.2.16",
+    "@angular-devkit/build-angular": "~0.803.29",
+    "@angular/cli": "8.3.29",
+    "@angular/compiler-cli": "8.2.14",
+    "@angular/language-service": "8.2.14",
     "@types/jasmine": "^2.8.12",
     "@types/jasminewd2": "^2.0.6",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^4.1.0",
@@ -58,6 +57,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": ">=3.1.1 <3.3"
+    "typescript": "3.5.3"
   }
 }
